break into a more classic MVC style
