+++ pkg/admin/controller_test.go
@@ -6,12 +6,11 @@ import (
 	"net"
 	"net/http"
 	"net/http/httptest"
-	"os"
-	"time"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/ginkgo/extensions/table"
 	. "github.com/onsi/gomega"
+	"github.com/sirupsen/logrus/hooks/test"
 
 	"github.com/pyroscope-io/pyroscope/pkg/admin"
 )
@@ -32,59 +31,59 @@ var _ = Describe("controller", func() {
 	// TODO
 	// since this talks with socket and stuff
 	// this can become an integration test
-	Context("/v1/apps", func() {
-		var (
-			httpC http.Client
-
-			socketAddr string
-		)
-
-		BeforeEach(func() {
-			file, err := ioutil.TempFile("", "pyroscope.sock")
-			must(err)
-			socketAddr = file.Name()
-
-			cfg := admin.Config{
-				SocketAddr: socketAddr,
-			}
-			svc := admin.NewService(mockAppsGetter{})
-			ctrl, err := admin.NewController(cfg, svc)
-			must(err)
-
-			go func() {
-				err := ctrl.Start()
-				if err != nil {
-					panic(err)
-				}
-			}()
-
-			httpC = newHttpClient(socketAddr)
-			// TODO
-			// for some reason it takes some time until server is ready
-			// we could add retries?
-			time.Sleep(time.Millisecond * 10)
-		})
-
-		AfterEach(func() {
-			os.Remove(socketAddr)
-		})
-
-		It("returns app names", func() {
-			req, err := httpC.Get("http://dummy/v1/apps")
-			Expect(err).To(BeNil())
-
-			body, err := ioutil.ReadAll(req.Body)
-			Expect(err).To(BeNil())
-
-			Expect(string(body)).To(Equal(`["app1","app2"]
-`))
-		})
-
-		It("only accepts GET requests", func() {
-
-			Expect(true).To(Equal(true))
-		})
-	})
+	//	Context("/v1/apps", func() {
+	//		var (
+	//			httpC http.Client
+	//
+	//			socketAddr string
+	//		)
+	//
+	//		BeforeEach(func() {
+	//			file, err := ioutil.TempFile("", "pyroscope.sock")
+	//			must(err)
+	//			socketAddr = file.Name()
+	//
+	//			cfg := admin.Config{
+	//				SocketAddr: socketAddr,
+	//			}
+	//			svc := admin.NewService(mockAppsGetter{})
+	//			ctrl, err := admin.NewController(cfg, svc)
+	//			must(err)
+	//
+	//			go func() {
+	//				err := ctrl.Start()
+	//				if err != nil {
+	//					panic(err)
+	//				}
+	//			}()
+	//
+	//			httpC = newHttpClient(socketAddr)
+	//			// TODO
+	//			// for some reason it takes some time until server is ready
+	//			// we could add retries?
+	//			time.Sleep(time.Millisecond * 10)
+	//		})
+	//
+	//		AfterEach(func() {
+	//			os.Remove(socketAddr)
+	//		})
+	//
+	//		It("returns app names", func() {
+	//			req, err := httpC.Get("http://dummy/v1/apps")
+	//			Expect(err).To(BeNil())
+	//
+	//			body, err := ioutil.ReadAll(req.Body)
+	//			Expect(err).To(BeNil())
+	//
+	//			Expect(string(body)).To(Equal(`["app1","app2"]
+	//`))
+	//		})
+	//
+	//		It("only accepts GET requests", func() {
+	//
+	//			Expect(true).To(Equal(true))
+	//		})
+	//	})
 
 	Context("/v1/apps", func() {
 		var svr *admin.AdminServer
@@ -92,10 +91,14 @@ var _ = Describe("controller", func() {
 
 		// create a server
 		BeforeEach(func() {
-			cfg := admin.Config{SocketAddr: "foo"}
+			// create a null logger, since we aren't interested
+			logger, _ := test.NewNullLogger()
+
+			cfg := admin.Config{SocketAddr: "foo", Log: logger}
 
 			svc := admin.NewService(mockAppsGetter{})
-			server, err := admin.NewServer(cfg, svc)
+			ctrl := admin.NewController(logger, svc)
+			server, err := admin.NewServer(cfg, ctrl)
 
 			must(err)
 			svr = server
